Create UserInput.py
This file will contain the communication with the user. It will be extended by further functions when we know the role of the government and know what options the user will have to let the government intervene in the economy as well as which numbers we will measure to give the user an level of equality we have in a specific economy.